chore: move hover related functions from utils
